feat: simplify interface to generalize rate limiter
BREAKING CHANGE: The New() function now returns a generic rate limiter to simplify implementation.
