+++ limiter/limiter.go
@@ -1,8 +1,15 @@
 package limiter
 
 import (
+	"fmt"
 	"time"
 )
+
+type Limiter struct {
+	RateLimiter
+	algorithm Algorithm
+	backend   Backend
+}
 
 type RateLimiter interface {
 	Evaluate(key string) (Calculation, error)
@@ -26,4 +33,48 @@ type Backend interface {
 type Configuration interface {
 	GetTokens() int
 	GetRefreshRate() time.Duration
+}
+
+func New(algorithm Algorithm, backend Backend) RateLimiter {
+	return &Limiter{
+		algorithm: algorithm,
+		backend:   backend,
+	}
+}
+
+func (rl Limiter) ShouldLimit(key string) (bool, error) {
+	evaluation, err := rl.Evaluate(key)
+	if err != nil {
+		return false, err
+	}
+
+	return evaluation.ShouldLimit, nil
+}
+
+func (rl Limiter) Evaluate(key string) (Calculation, error) {
+	now := time.Now()
+	currentWindow := now.Truncate(rl.algorithm.GetRefreshRate())
+	nextWindow := currentWindow.Add(rl.algorithm.GetRefreshRate())
+
+	keyPrefix := currentWindow.Format(time.Stamp)
+	redisKey := fmt.Sprintf("%s %s", keyPrefix, key)
+
+	tokens, err := rl.backend.Get(redisKey)
+	if err != nil {
+		panic(err)
+	}
+	defer func(newTokens int) {
+		err := rl.backend.Set(redisKey, newTokens, rl.algorithm.GetRefreshRate())
+		if err != nil {
+			panic(err)
+		}
+	}(tokens + 1)
+
+	shouldLimit := tokens >= rl.algorithm.GetTokens()
+	retryAfter := nextWindow.Sub(now)
+
+	return Calculation{
+		ShouldLimit: shouldLimit,
+		RetryAfter:  retryAfter,
+	}, nil
 }
